move errorHandler wrap method to own file
